# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Finance Tracker API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type HoldingType = (typeof HoldingType)[keyof typeof HoldingType];

export const HoldingType = {
  stock: "stock",
  gold: "gold",
} as const;

export interface Holding {
  id: number;
  type: HoldingType;
  symbol: string;
  quantity: number;
  createdAt: string;
  updatedAt: string;
}

export type CreateHoldingBodyType =
  (typeof CreateHoldingBodyType)[keyof typeof CreateHoldingBodyType];

export const CreateHoldingBodyType = {
  stock: "stock",
  gold: "gold",
} as const;

export interface CreateHoldingBody {
  type: CreateHoldingBodyType;
  symbol: string;
  quantity: number;
}

export interface UpdateHoldingBody {
  quantity: number;
}

export type PriceType = (typeof PriceType)[keyof typeof PriceType];

export const PriceType = {
  stock: "stock",
  gold: "gold",
} as const;

export interface Price {
  id: number;
  type: PriceType;
  symbol: string;
  price: number;
  /** @nullable */
  change: number | null;
  /** @nullable */
  changePercent: number | null;
  fetchedAt: string;
}

export interface RefreshResult {
  success: boolean;
  updated: number;
  message: string;
}

export interface Snapshot {
  id: number;
  totalValue: number;
  stockValue: number;
  goldValue: number;
  snapshotAt: string;
}

export type HoldingWithValueType =
  (typeof HoldingWithValueType)[keyof typeof HoldingWithValueType];

export const HoldingWithValueType = {
  stock: "stock",
  gold: "gold",
} as const;

export interface HoldingWithValue {
  id: number;
  type: HoldingWithValueType;
  symbol: string;
  quantity: number;
  /** @nullable */
  currentPrice: number | null;
  /** @nullable */
  currentValue: number | null;
  /** @nullable */
  change: number | null;
  /** @nullable */
  changePercent: number | null;
}

export interface PortfolioSummary {
  totalValue: number;
  stockValue: number;
  goldValue: number;
  /** @nullable */
  lastUpdated: string | null;
  holdings: HoldingWithValue[];
}