# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Finance Tracker API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all holdings
 */
export const ListHoldingsResponseItem = zod.object({
  id: zod.number(),
  type: zod.enum(["stock", "gold"]),
  symbol: zod.string(),
  quantity: zod.number(),
  createdAt: zod.date(),
  updatedAt: zod.date(),
});
export const ListHoldingsResponse = zod.array(ListHoldingsResponseItem);

/**
 * @summary Create a new holding
 */
export const CreateHoldingBody = zod.object({
  type: zod.enum(["stock", "gold"]),
  symbol: zod.string(),
  quantity: zod.number(),
});

/**
 * @summary Update a holding
 */
export const UpdateHoldingParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateHoldingBody = zod.object({
  quantity: zod.number(),
});

export const UpdateHoldingResponse = zod.object({
  id: zod.number(),
  type: zod.enum(["stock", "gold"]),
  symbol: zod.string(),
  quantity: zod.number(),
  createdAt: zod.date(),
  updatedAt: zod.date(),
});

/**
 * @summary Delete a holding
 */
export const DeleteHoldingParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Get latest prices for all tracked symbols
 */
export const GetLatestPricesResponseItem = zod.object({
  id: zod.number(),
  type: zod.enum(["stock", "gold"]),
  symbol: zod.string(),
  price: zod.number(),
  change: zod.number().nullable(),
  changePercent: zod.number().nullable(),
  fetchedAt: zod.date(),
});
export const GetLatestPricesResponse = zod.array(GetLatestPricesResponseItem);

/**
 * @summary Trigger a price refresh
 */
export const RefreshPricesResponse = zod.object({
  success: zod.boolean(),
  updated: zod.number(),
  message: zod.string(),
});

/**
 * @summary Get portfolio value snapshots for the last 7 days
 */
export const ListSnapshotsResponseItem = zod.object({
  id: zod.number(),
  totalValue: zod.number(),
  stockValue: zod.number(),
  goldValue: zod.number(),
  snapshotAt: zod.date(),
});
export const ListSnapshotsResponse = zod.array(ListSnapshotsResponseItem);

/**
 * @summary Get total portfolio value with current prices
 */
export const GetPortfolioSummaryResponse = zod.object({
  totalValue: zod.number(),
  stockValue: zod.number(),
  goldValue: zod.number(),
  lastUpdated: zod.date().nullable(),
  holdings: zod.array(
    zod.object({
      id: zod.number(),
      type: zod.enum(["stock", "gold"]),
      symbol: zod.string(),
      quantity: zod.number(),
      currentPrice: zod.number().nullable(),
      currentValue: zod.number().nullable(),
      change: zod.number().nullable(),
      changePercent: zod.number().nullable(),
    }),
  ),
});