# lib/api-zod/src/generated/types/createHoldingBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Finance Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateHoldingBodyType } from "./createHoldingBodyType";

export interface CreateHoldingBody {
  type: CreateHoldingBodyType;
  symbol: string;
  quantity: number;
}