# lib/api-zod/src/generated/types/holding.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Finance Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { HoldingType } from "./holdingType";

export interface Holding {
  id: number;
  type: HoldingType;
  symbol: string;
  quantity: number;
  createdAt: Date;
  updatedAt: Date;
}