# lib/api-zod/src/generated/types/holdingWithValue.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Finance Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { HoldingWithValueType } from "./holdingWithValueType";

export interface HoldingWithValue {
  id: number;
  type: HoldingWithValueType;
  symbol: string;
  quantity: number;
  /** @nullable */
  currentPrice: number | null;
  /** @nullable */
  currentValue: number | null;
  /** @nullable */
  change: number | null;
  /** @nullable */
  changePercent: number | null;
}