# lib/api-zod/src/generated/types/price.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Finance Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { PriceType } from "./priceType";

export interface Price {
  id: number;
  type: PriceType;
  symbol: string;
  price: number;
  /** @nullable */
  change: number | null;
  /** @nullable */
  changePercent: number | null;
  fetchedAt: Date;
}