# lib/api-zod/src/generated/types/snapshot.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Finance Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface Snapshot {
  id: number;
  totalValue: number;
  stockValue: number;
  goldValue: number;
  snapshotAt: Date;
}